refactor: suggest changes for #649
