+++ i18n/english.yml
@@ -245,12 +245,15 @@ components:
       label: Delete records from %tablePlaceholder%
       name: Delete records transformation
     NormalizeFieldTransformation:
+      filePlaceholder: Choose file/table to normalize
       label: Normalize field
       name: Normalize field transformation
     ReplaceFileFromStringTransformation:
+      filePlaceholder: Choose the file/table to replace
       label: Replace %tablePlaceholder% from %filePlaceholder%
       name: Replace file from string transformation
     ReplaceFileFromVersionTransformation:
+      filePlaceholder: Choose the file/table to replace
       label: Replace %tablePlaceholder% from %versionPlaceholder%
       name: Replace file from version transformation
   FeedVersionNavigator:
