+++ lib/manager/components/transform/FeedTransformation.js
@@ -6,7 +6,7 @@ import {Button} from 'react-bootstrap'
 import Select from 'react-select'
 
 import {getGtfsSpec, getGtfsPlusSpec, isModuleEnabled} from '../../../common/util/config'
-import {getTransformationName} from '../../util/transform'
+import {getTransformationName, getTransformationPlaceholder} from '../../util/transform'
 import NormalizeField from './NormalizeField'
 import ReplaceFileFromString from './ReplaceFileFromString'
 import ReplaceFileFromVersion from './ReplaceFileFromVersion'
@@ -114,7 +114,7 @@ export default class FeedTransformation extends Component<Props, {errors: Array<
           <Select
             style={{margin: '10px 0px', width: '230px'}}
             clearable={false}
-            placeholder='Choose the file/table to replace'
+            placeholder={getTransformationPlaceholder(transformationType, 'filePlaceholder')}
             options={tables.map(table => ({value: table.name.split('.txt')[0], label: table.name}))}
             value={transformation.table}
             onChange={this._onSelectTable} />
