+++ lib/manager/components/transform/NormalizeField.js
@@ -19,9 +19,9 @@ import SubstitutionRow from './SubstitutionRow'
 import type {
   NormalizeFieldFields,
   ReactSelectOption,
-  Substitution
+  Substitution,
+  TransformProps
 } from '../../../types'
-import type { TransformProps } from './Transform'
 
 /**
  * Extract just the subset of fields to build the state variable.
@@ -44,7 +44,7 @@ function extractStateFields (transformation: NormalizeFieldFields): NormalizeFie
 }
 
 /**
- * Component that renders input fields for the NoramlizeFieldTransformation.
+ * Component that renders input fields for the NormalizeFieldTransformation.
  */
 export default class NormalizeField extends Component<TransformProps<NormalizeFieldFields>, NormalizeFieldFields> {
   constructor (props: TransformProps<NormalizeFieldFields>) {
