+++ lib/manager/components/transform/ReplaceFileFromString.js
@@ -3,8 +3,7 @@
 import React, { Component } from 'react'
 import { Button } from 'react-bootstrap'
 
-import type { ReplaceFileFromStringFields } from '../../../types'
-import type { TransformProps } from './Transform'
+import type { ReplaceFileFromStringFields, TransformProps } from '../../../types'
 
 /**
  * Component that renders fields for ReplaceFileFromStringTransformation.
