+++ lib/manager/components/transform/ReplaceFileFromVersion.js
@@ -4,7 +4,7 @@ import React, { Component } from 'react'
 
 import VersionSelectorDropdown from '../version/VersionSelectorDropdown'
 
-import type { TransformProps } from './Transform'
+import type { TransformProps } from '../../../types'
 
 /**
  * Component that renders fields for ReplaceFileFromVersionTransformation.
