+++ lib/manager/util/transform.js
@@ -54,3 +54,17 @@ export function getTransformationName (
   })
   return label
 }
+
+export function getTransformationPlaceholder (
+  type: string,
+  placeholderId: string
+) {
+  // Get the base component messages.
+  const messages = getComponentMessages('FeedTransformationDescriptions')
+  const messageId = `${type}.${placeholderId}`
+  const message = messages(messageId)
+  // If message does not exist for component, fall back on general message.
+  return message.indexOf(messageId) !== -1
+    ? messages(`general.${placeholderId}`)
+    : message
+}
