+++ lib/types/index.js
@@ -319,6 +319,14 @@ export type FeedTransformationBase = {
   table: ?string
 }
 
+export type TransformProps<StateType> = {
+  feedSource: Feed,
+  index: number,
+  onSave: (StateType, number) => void,
+  onValidationErrors: (Array<string>) => void,
+  transformation: FeedTransformationBase & StateType
+}
+
 export type DeleteRecordsTransformation =
   & {typeName: 'DeleteRecordsTransformation'}
   & FeedTransformationBase
